% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_defaultPanels}
\alias{add_defaultPanels}
\title{Add Default Gene Panels to iSensor Object}
\usage{
add_defaultPanels(iSensor_obj)
}
\arguments{
\item{iSensor_obj}{An iSensor object to which the default gene panels will be added.}
}
\value{
Updated iSensor object with default gene panels included.
}
\description{
Loads default gene panels from the package's extdata directory and adds them 
to the specified iSensor object.
}
\section{Details}{

This function retrieves gene panel directories from the package's extdata folder, 
reads the panel files, and adds them to the iSensor object as new panel sets. 
Each folder in the directory corresponds to a different hormone panel.
}

\examples{
\dontrun{
# Example of adding default panels to an iSensor object
iSensor_obj <- add_defaultPanels(iSensor_obj)
}

}
\seealso{
\code{\link{iSensor_pipeline}}, \code{\link{set_iSensor_workPanel}}, 
\code{\link{add_iSensor_panelSet}}, \code{\link{read_genePanels}}
}
