% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_iSensor_labels}
\alias{add_iSensor_labels}
\title{Add Sample Labels to iSensor Object}
\usage{
add_iSensor_labels(iSensor_obj, labels)
}
\arguments{
\item{iSensor_obj}{An iSensor object to which the sample labels will be added.}

\item{labels}{A character vector containing labels for each sample in the expression data.}
}
\value{
Updated iSensor object with the new sample labels included.
}
\description{
Assigns sample labels to the specified iSensor object, linking them to the expression data.
}
\section{Details}{

This function checks if the length of the provided labels matches the number of samples 
in the expression data. If the lengths match, the labels are assigned to the iSensor object, 
allowing for easier identification of samples during analysis.
}

\examples{
\dontrun{
# Example of adding sample labels to an iSensor object
iSensor_obj <- add_iSensor_labels(iSensor_obj, c("Sample1", "Sample2", "Sample3"))
}

}
\seealso{
\code{\link{iSensor_pipeline}}, \code{\link{add_iSensor_data}}, 
\code{\link{set_iSensor_workPanel}}
}
