% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_iSensor_panelSet}
\alias{add_iSensor_panelSet}
\title{Add Panel Set to iSensor Object}
\usage{
add_iSensor_panelSet(iSensor_obj, panelsSet, setName = NULL)
}
\arguments{
\item{iSensor_obj}{An iSensor object (created by `create_iSensor`)}

\item{panelsSet}{List of gene panels to add (either a single panel or multiple panels)}

\item{setName}{Optional name for the panel set (autogenerated if NULL)}
}
\value{
The modified iSensor object with the new panel set added to `genePanelSets`
}
\description{
Adds a new set of gene panels to an iSensor object. The panel set can be either
a single panel or multiple panels combined into a named set.
}
\examples{
# Create iSensor object
iSensor_obj <- create_iSensor(data = matrix(rnorm(1000), nrow = 100))

# Add a new panel set
new_panels <- list(
  "CellCycle" = c("CDK1", "CDK2", "CCNB1"),
  "Apoptosis" = c("BAX", "BCL2", "CASP3")
)
iSensor_obj <- add_iSensor_panelSet(iSensor_obj, new_panels, setName = "MyPanels")

}
