% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_iSensor_random_panel}
\alias{add_iSensor_random_panel}
\title{Add Random Gene Panels to iSensor Object}
\usage{
add_iSensor_random_panel(iSensor_obj, randNum, randSize, majortrend = TRUE)
}
\arguments{
\item{iSensor_obj}{An iSensor object to which the random gene panels will be added.}

\item{randNum}{Integer specifying the number of random panels to generate.}

\item{randSize}{Numeric vector specifying the sizes of the random panels. The length 
of this vector must match `randNum`.}

\item{majortrend}{Logical indicating whether to include a major trend panel based on 
all features. Default is TRUE.}
}
\value{
Updated iSensor object with new random gene panels included.
}
\description{
Generates random gene panels based on the expression data in the iSensor object and 
adds them to the specified iSensor object.
}
\section{Details}{

This function checks if the expression data in the iSensor object is available. It then 
generates random gene panels of specified sizes and adds them to the iSensor object. 
A major trend panel, containing all features, is also included by default.
}

\examples{
\dontrun{
# Example of adding random gene panels to an iSensor object
iSensor_obj <- add_iSensor_random_panel(iSensor_obj, randNum = 3, randSize = c(100, 200, 300))
}

}
\seealso{
\code{\link{iSensor_pipeline}}, \code{\link{add_iSensor_workPanel}}, 
\code{\link{add_defaultPanels}}
}
