% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_iSensor_workPanel}
\alias{add_iSensor_workPanel}
\title{Add Gene Panels to iSensor Object}
\usage{
add_iSensor_workPanel(iSensor_obj, panels, drop = FALSE)
}
\arguments{
\item{iSensor_obj}{An iSensor object to which the gene panels will be added.}

\item{panels}{A list of gene panels to be added or updated in the iSensor object.}

\item{drop}{Logical indicating whether to drop existing panels that are not included in the new panels. Default is FALSE.}
}
\value{
Updated iSensor object with the new and updated gene panels included.
}
\description{
Incorporates additional gene panels into the specified iSensor object, updating existing panels 
and adding new ones as necessary.
}
\section{Details}{

This function checks for existing gene panels in the iSensor object and updates them 
with the new panels provided. If a panel already exists, it will be replaced; otherwise, 
it will be added to the iSensor object. The `drop` parameter controls whether panels 
not included in the new list should be removed.
}

\examples{
\dontrun{
# Example of adding gene panels to an iSensor object
new_panels <- list(panel1 = gene_data1, panel2 = gene_data2)
iSensor_obj <- add_iSensor_workPanel(iSensor_obj, new_panels)
}

}
\seealso{
\code{\link{iSensor_pipeline}}, \code{\link{set_iSensor_workPanel}}, 
\code{\link{add_defaultPanels}}
}
