% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{read_genePanels}
\alias{read_genePanels}
\title{Read Gene Panels from Files}
\usage{
read_genePanels(panelsDir, panelsFiles, panelsNames = NULL)
}
\arguments{
\item{panelsDir}{Character string specifying the directory containing the gene panel files.}

\item{panelsFiles}{Character vector of file names to be read from the specified directory.}

\item{panelsNames}{Optional character vector of names to assign to the gene panels. 
If NULL, names will be derived from the file names. Default is NULL.}
}
\value{
A list of data frames, each containing a column of gene names corresponding to 
        the specified panel files.
}
\description{
Reads gene panel data from specified files in a directory and returns a list of data frames 
containing gene information.
}
\section{Details}{

This function reads gene panel files from the specified directory, creating a data frame 
for each file that contains the gene names. If panel names are provided, they will be used 
as the names of the list elements; otherwise, names will be generated from the file names 
by removing the ".txt" extension.
}

\examples{
\dontrun{
# Example of reading gene panels from a directory
gene_panels <- read_genePanels("path/to/panels", c("panel1.txt", "panel2.txt"))
}

}
\seealso{
\code{\link{add_defaultPanels}}, \code{\link{add_iSensor_panelSet}}, 
\code{\link{set_iSensor_workPanel}}
}
