% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_iSensor_data}
\alias{add_iSensor_data}
\title{Add Expression Data to iSensor Object}
\usage{
add_iSensor_data(iSensor_obj, data)
}
\arguments{
\item{iSensor_obj}{An iSensor object to which the expression data will be added.}

\item{data}{A matrix or data frame containing the expression data to be assigned.}
}
\value{
Updated iSensor object with the new expression data included.
}
\description{
Assigns expression data to the specified iSensor object for further analysis.
}
\section{Details}{

This function updates the expression data slot of the iSensor object with the 
provided data, allowing for subsequent analysis using the iSensor framework.
}

\examples{
\dontrun{
# Example of adding expression data to an iSensor object
iSensor_obj <- add_iSensor_data(iSensor_obj, expression_data_matrix)
}

}
\seealso{
\code{\link{iSensor_pipeline}}, \code{\link{set_iSensor_workPanel}}, 
\code{\link{add_defaultPanels}}
}
