% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_iSensor_panelSet}
\alias{add_iSensor_panelSet}
\title{Add Gene Panel Set to iSensor Object}
\usage{
add_iSensor_panelSet(iSensor_obj, panelsSet, setName = NULL)
}
\arguments{
\item{iSensor_obj}{An iSensor object to which the gene panel set will be added.}

\item{panelsSet}{A list of gene panels to be included in the new panel set.}

\item{setName}{Optional character string specifying the name of the new panel set. 
If NULL, a default name will be generated. Default is NULL.}
}
\value{
Updated iSensor object with the new gene panel set included.
}
\description{
Adds a new set of gene panels to the specified iSensor object, allowing for organized 
management of multiple panel sets.
}
\section{Details}{

This function allows users to add a new set of gene panels to the iSensor object. 
If a set name is not provided, a default name is generated based on the current number 
of panel sets. The function ensures that the new panel set is properly integrated into 
the existing structure of the iSensor object.
}

\examples{
\dontrun{
# Example of adding a gene panel set to an iSensor object
new_panel_set <- list(panel1 = gene_data1, panel2 = gene_data2)
iSensor_obj <- add_iSensor_panelSet(iSensor_obj, new_panel_set, setName = "CustomPanelSet")
}

}
\seealso{
\code{\link{iSensor_pipeline}}, \code{\link{add_iSensor_workPanel}}, 
\code{\link{add_defaultPanels}}
}
