% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{add_iSensor_random_panel}
\alias{add_iSensor_random_panel}
\title{Add Random Gene Panels to iSensor Object}
\usage{
add_iSensor_random_panel(iSensor_obj, randNum, randSize, majortrend = TRUE)
}
\arguments{
\item{iSensor_obj}{An iSensor object created by `create_iSensor`}

\item{randNum}{Number of random panels to generate (must match length of `randSize`)}

\item{randSize}{Vector specifying sizes for each random panel}

\item{majortrend}{Logical indicating whether to include a panel with all genes (default: TRUE)}
}
\value{
The modified iSensor object with added random panels
}
\description{
Generates and adds random gene panels to an iSensor object. Can optionally include
a "majortrend" panel containing all genes. Each random panel samples genes without replacement.
}
\examples{
# Create iSensor object with 1000 genes
iSensor_obj <- create_iSensor(data = matrix(rnorm(10000), nrow = 1000))

# Add 3 random panels of different sizes
iSensor_obj <- add_iSensor_random_panel(iSensor_obj, 
                                      randNum = 3,
                                      randSize = c(50, 100, 200))

}
