% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor-package.R
\docType{package}
\name{iSensors}
\alias{iSensors}
\alias{iSensors-package}
\title{iSensors: Single-Cell Gene Panel Analysis Package}
\description{
The iSensors package provides tools for analyzing single-cell gene expression data 
using predefined and custom gene panels. It allows users to compute signaling activity 
scores, manage gene panels, and perform various analyses on single-cell datasets.
}
\seealso{
\code{\link{LoadSensors}}, \code{\link{CalcSensors}}, \code{\link{iSensorsCisTransPanelCreate}}, \code{\link{iSensorsTransPanelCreate}}
}
\author{
Maxim Rybakov [aut, cre], 
Elena Zemlyanskaya [aut], 
Victoria Mironova [aut], 
Vladislav Dolgikh [aut]
}
\keyword{package}
