% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_creation.R
\name{iSensorsCisTransPanelCreate}
\alias{iSensorsCisTransPanelCreate}
\title{The iSensorsCisTransPanelCreate function for creating CisTrans panels based on motif recognition in promoters.}
\usage{
iSensorsCisTransPanelCreate(
  panel_name,
  species,
  promoters_set,
  ppm,
  deg_list,
  min_dataset_number,
  trivial_names_file,
  panel_type,
  transcriptomes_info
)
}
\arguments{
\item{panel_name}{The name of the panel to create. The function will create a GenePanel object with this name and save it as rda file in the user_panels directory.}

\item{species}{Name of species for which the trans panel is created.}

\item{promoters_set}{Path to a FASTA file containing promoter sequences.}

\item{ppm}{Path to a positional probability matrix file in Homer format.}

\item{deg_list}{Path to a file containing logFC and adjusted p-value information obtained from a set of transcriptomic experiments that determine the differential expression status of a gene.}

\item{min_dataset_number}{The minimum number of transcriptomics experiments in which a gene is differentially expressed to be considered differentially expressed in the cis-trans panel. For example, if min_dataset_number = 1, then a gene must be differentially expressed in at least one experiment to be considered differentially expressed in the CisTrans panel.}

\item{trivial_names_file}{List of trivial gene names as a .txt file. (optional)}

\item{panel_type}{Type of panel. Could be cis, UP and DOWN.}

\item{transcriptomes_info}{Information about the used transcriptomes, written by the user.}
}
\description{
The iSensorsCisTransPanelCreate function performs the recognition of binding sites in promoters using positional weight matrices and generates a GenePanel object for a CisTrans type panel. 
Format for CisTrans panels is compatible with iSensors R package. 
The function receives (1) a set of promoters in FASTA format, (2) a positional probability matrix in Homer format, (3) a list of RNA-seq experiments with logFC and adjusted p-values for each gene (optional) and (4) the the corresponding trivial gene names in txt format (optional).
The function generates a list of three items. 
The first is a list genes containing gene IDs. 
The second is a data frame gene_metadata containing gene IDs, recognized sites, coordinates of each site in the genome, location (forward or reverse strand) and distance of each site relative to the TSS, and the corresponding short and full gene names. 
The third is a data frame panel_metadata containing the information about the species, gene panel type, gene panel description, creation date. 
The function writes the panel as an rda file in the user_panels directory (and creates it if this directory doesn't exist).
}
\examples{
\dontrun{
# Arabidopsis thaliana, cis panel.

iSensorsCisTransPanelCreate(panel_name = 'cis_panel',  
species = 'Arabidopsis thaliana',  
promoters_set = 'At_TAIR10_promoters.fas', 
ppm = 'ARF1.txt', panel_type = 'cis',  
trivial_names_file = 'at_trivial.txt')

# Arabidopsis thaliana, UP cis-trans panel.

iSensorsCisTransPanelCreate(panel_name = 'cis-trans_panel',  
species = 'Arabidopsis thaliana', 
promoters_set = 'At_TAIR10_promoters.fas', 
ppm = 'ARF1.txt', 
deg_list = 'auxin_degs.txt', 
min_dataset_number = 1, 
panel_type = 'UP', 
transcriptomes_info = 'Auxin 1h, auxin 4h')
}

}
