% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_creation.R
\name{iSensorsTransPanelCreate}
\alias{iSensorsTransPanelCreate}
\title{The iSensorsTransPanelCreate function generates a GenePanel object for a Trans type gene panel compatible with iSensors R package.}
\usage{
iSensorsTransPanelCreate(
  panel_name,
  species,
  gene_list,
  trivial_names_file,
  panel_description
)
}
\arguments{
\item{panel_name}{The name of the panel to create. The function will create a GenePanel object with this name and save it as rda file in the user_panels directory.}

\item{species}{Name of species for which the trans panel is created.}

\item{gene_list}{Vector of genes. Can be supplied as a vector c('AT1G01010', 'AT1G01030', 'AT1G01040'), or as a .txt file.}

\item{trivial_names_file}{List of trivial gene names as a .txt file. (optional)}

\item{panel_description}{Description of the panel}
}
\description{
The function receives a list of gene IDs as input (either as a vector or as a txt file), and the the corresponding trivial gene names in txt format (optional).
The function generates a list of three items. The first is a list genes containing gene IDs. 
The second is a data frame gene_metadata containing gene IDs and the corresponding short and full gene names. 
The third is a data frame panel_metadata containing the information about the species, gene panel type, gene panel description, creation date. 
The function writes the panel to an object in the current environment and saves it as an rda file in the current working directory.
}
\examples{
\dontrun{
# Arabidopsis thaliana, input as vector.
iSensorsTransPanelCreate(panel_name = 'trans_panel', 
species = 'Arabidopsis thaliana', 
gene_list = c('AT1G01010', 'AT1G01030', 'AT1G01040'), 
panel_description = 'This is an example of trans panel')

# Arabidopsis thaliana, input as txt-file.

iSensorsTransPanelCreate(panel_name = 'trans_panel', 
species = 'Arabidopsis thaliana', 
gene_list = 'gene_list.txt', 
panel_description = 'This is an example of trans panel')

# Arabidopsis thaliana, input as txt-file, with trivial names list.

iSensorsTransPanelCreate(panel_name = 'trans_panel', 
species = 'Arabidopsis thaliana', 
gene_list = 'gene_list.txt', 
trivial_names_file = 'Arabidopsis_trivial_names_example.txt', 
panel_description = 'This is an example of trans panel')
}

}
