% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSensor.R
\name{set_iSensor_workPanel}
\alias{set_iSensor_workPanel}
\title{Set Working iSensor Gene Panel}
\usage{
set_iSensor_workPanel(iSensor_obj, panelSet = "Auxin")
}
\arguments{
\item{iSensor_obj}{An iSensor object containing gene panel sets.}

\item{panelSet}{Character string specifying the name of the gene panel set to use. 
Default is 'Auxin'.}
}
\value{
Updated iSensor object with the selected gene panel set assigned.
}
\description{
Configures the iSensor object to use a specified gene panel set for analysis.
}
\section{Details}{

This function checks if the specified panel set exists in the iSensor object. 
If it does, the function assigns the corresponding gene panels to the iSensor object 
for subsequent analysis.
}

\examples{
\dontrun{
# Example of setting a working panel
iSensor_obj <- set_iSensor_workPanel(iSensor_obj, panelSet = 'Cytokinin')
}

}
\seealso{
\code{\link{iSensor_pipeline}}, \code{\link{create_iSensor}}, 
\code{\link{add_iSensor_random_panel}}, \code{\link{iSensor_signal}}
}
